refactor: use `enum ffa_error` for errors
Convert variables intended to hold an FF-A error to `enum ffa_error`
from `uint23_t`

+++ inc/hf/ffa_internal.h
@@ -19,8 +19,7 @@
 
 static inline struct ffa_value ffa_error(enum ffa_error error_code)
 {
-	return (struct ffa_value){.func = FFA_ERROR_32,
-				  .arg2 = (uint32_t)error_code};
+	return (struct ffa_value){.func = FFA_ERROR_32, .arg2 = error_code};
 }
 
 struct ffa_value ffa_msg_recv_return(const struct vm *receiver);
