+++ test/vmapi/ffa_secure_partitions/notifications.c
@@ -266,7 +266,7 @@ TEST(ffa_notifications, fail_if_mbz_set_in_notification_set)
 				    FFA_NOTIFICATION_MASK(1));
 	EXPECT_EQ(res.func, FFA_MSG_SEND_DIRECT_RESP_32);
 	EXPECT_EQ(sp_resp(res), SP_ERROR);
-	EXPECT_EQ(sp_resp_value(res), FFA_INVALID_PARAMETERS);
+	EXPECT_EQ((int32_t)sp_resp_value(res), FFA_INVALID_PARAMETERS);
 }
 
 /**
@@ -290,7 +290,7 @@ TEST(ffa_notifications, fail_if_global_notif_vcpu_not_zero)
 				    FFA_NOTIFICATION_MASK(1));
 	EXPECT_EQ(res.func, FFA_MSG_SEND_DIRECT_RESP_32);
 	EXPECT_EQ(sp_resp(res), SP_ERROR);
-	EXPECT_EQ(sp_resp_value(res), FFA_INVALID_PARAMETERS);
+	EXPECT_EQ((int32_t)sp_resp_value(res), FFA_INVALID_PARAMETERS);
 }
 
 TEST(ffa_notifications, fail_if_global_notif_set_as_per_vcpu)
@@ -310,7 +310,7 @@ TEST(ffa_notifications, fail_if_global_notif_set_as_per_vcpu)
 				    FFA_NOTIFICATION_MASK(1));
 	EXPECT_EQ(res.func, FFA_MSG_SEND_DIRECT_RESP_32);
 	EXPECT_EQ(sp_resp(res), SP_ERROR);
-	EXPECT_EQ(sp_resp_value(res), FFA_INVALID_PARAMETERS);
+	EXPECT_EQ((int32_t)sp_resp_value(res), FFA_INVALID_PARAMETERS);
 }
 
 TEST(ffa_notifications, fail_if_per_vcpu_notif_set_as_global)
@@ -330,7 +330,7 @@ TEST(ffa_notifications, fail_if_per_vcpu_notif_set_as_global)
 				    FFA_NOTIFICATION_MASK(1));
 	EXPECT_EQ(res.func, FFA_MSG_SEND_DIRECT_RESP_32);
 	EXPECT_EQ(sp_resp(res), SP_ERROR);
-	EXPECT_EQ(sp_resp_value(res), FFA_INVALID_PARAMETERS);
+	EXPECT_EQ((int32_t)sp_resp_value(res), FFA_INVALID_PARAMETERS);
 }
 
 TEST(ffa_notifications, fail_if_mbz_set_in_notifications_bind)
